fix: accidental clones within another repo
For years, this script has sometimes cloned a bunch of repos into the
base folder of another repo. The reason is that if it encounters any
error (networking, a remote that no longer exists, etc), it wasn't
popping back out of that repo directory before proceeding with the rest
of the repos. This fixes that.